requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.4
PyJWT==2.8.0
cryptography==41.0.3
python-jose==3.3.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
redis==4.5.5
requests==2.31.0
